requirements.txt
amt==0.8.0
kopf==1.36.0
kubernetes==25.3.0
PyYAML==6.0
